+++ PokemonUnity.Application/Overworld/Map.cs
@@ -26,7 +26,7 @@ namespace PokemonUnity
 			public int fog_hue				            { get; set; }     // fog hue
 			public float fog_opacity				    { get; set; }     // fog opacity level
 			public int fog_blend_type				    { get; set; }     // fog blending method
-			public float fog_zoom				        { get; set; }     // fog zoom rate
+			public int fog_zoom				        { get; set; }     // fog zoom rate
 			public float fog_sx				            { get; set; }     // fog sx
 			public float fog_sy				            { get; set; }     // fog sy
 			public string battleback_name			    { get; set; }     // battleback file name
